backend-app-api: move rootHttpRouter into subfolder
